Refactor services module
