+++ Causal_Web/engine/services/__init__.py
@@ -1,0 +1,25 @@
+"""Service modules for the engine package."""
+
+from .node_services import (
+    NodeTickService,
+    EdgePropagationService,
+    NodeTickDecisionService,
+)
+from .sim_services import (
+    NodeMetricsResultService,
+    NodeMetricsService,
+    GraphLoadService,
+    BridgeApplyService,
+    GlobalDiagnosticsService,
+)
+
+__all__ = [
+    "NodeTickService",
+    "EdgePropagationService",
+    "NodeTickDecisionService",
+    "NodeMetricsResultService",
+    "NodeMetricsService",
+    "GraphLoadService",
+    "BridgeApplyService",
+    "GlobalDiagnosticsService",
+]
